requirements.txt
Flask>=2.3
Flask-Login>=0.6
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
openpyxl>=3.1
gunicorn>=21
PyJWT>=2.9
pypinyin>=0.55
